[refactor] : PostLayout의 Header, Content 파일 분리
